Refactor weather data handling and update application structure
- Changed the main entry point to use App.tsx instead of main.tsx.
- Added @tanstack/react-query-devtools for enhanced query debugging.
- Introduced useWeather hook for fetching weather data.
- Moved GetWeather function to a new service file for better organization.
- Updated WeatherWidget to utilize the new useWeather hook.
- Removed obsolete GetWeather component.
- Created TestRender component for rendering admin layout.
