+++ frontend/package.json
@@ -17,6 +17,7 @@
     "@mui/material": "^7.3.8",
     "@mui/x-date-pickers": "^8.27.2",
     "@tanstack/react-query": "^5.90.21",
+    "@tanstack/react-query-devtools": "^5.91.3",
     "dayjs": "^1.11.19",
     "react": "^19.2.0",
     "react-dnd": "^16.0.1",
