+++ frontend/src/App.tsx
@@ -4,34 +4,17 @@ import { AdminLayout } from "./components/admin/layout/AdminLayout";
 import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 import { Outlet } from "react-router";
-import Watch from "./components/base/watch";
+import { createRoot } from "react-dom/client";
+import Main from "./main";
 
-function App() {
-  return ( 
-    <>
-      <h1>DashyBoard</h1>
-      <Watch location="Stockholm" timeZone="UTC"></Watch>
+createRoot(document.getElementById("root")!).render(
+  <>
+    <Main />
 
-      <nav>
-        <ul>
-          <li>
-            <Link to="/">Admin</Link>
-          </li>
-          <li>
-            <Link to="/room/123">Room -default</Link>
-          </li>
-          <li>
-            <Link to="/mirror">Spegel</Link>
-          </li>
-        </ul>
-      </nav>
-      <LocalizationProvider dateAdapter={AdapterDayjs}>
-        <AdminLayout>
-          <Outlet />
-        </AdminLayout>
-      </LocalizationProvider>
-    </>
-  );
-}
-
-export default App;
+    {/* <LocalizationProvider dateAdapter={AdapterDayjs}>
+      <AdminLayout>
+        <Outlet />
+      </AdminLayout>
+    </LocalizationProvider> */}
+  </>,
+);
