+++ frontend/src/components/admin/RenderAdminLayout.tsx
@@ -1,0 +1,12 @@
+import { AdminLayout } from "./layout/AdminLayout";
+import { Outlet } from "react-router";
+
+function TestRender() {
+  return (
+    <AdminLayout>
+      <Outlet />
+    </AdminLayout>
+  );
+}
+
+export default TestRender;
