+++ frontend/src/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useWeather } from "./useWeather";
+
